change the parameter type of x10.io.Reader from Int to Long
git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@26735 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10.runtime/src-java/x10/core/io/InputStream.java
@@ -117,7 +117,7 @@ public class InputStream extends Ref {
         }
     }
 
-    public void skip(int n) {
+    public void skip(long n) {
         try {
             stream.skip(n);
         } catch (java.io.IOException e) {
